# mgr/server/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  clearbill_mgr_server_api_vo.CreateUserReq:
    properties:
      displayName:
        maxLength: 128
        type: string
      role:
        maxLength: 64
        type: string
      status:
        maxLength: 32
        type: string
      tenantId:
        type: integer
      username:
        maxLength: 64
        type: string
    required:
    - displayName
    - role
    - username
    type: object
...
  clearbill_mgr_server_api_vo.ResetPasswordReq:
    properties:
      newPassword:
        maxLength: 128
        minLength: 6
        type: string
    required:
    - newPassword
    type: object
  clearbill_mgr_server_api_vo.ResponseResult:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  clearbill_mgr_server_api_vo.UpdateUserReq:
    properties:
      displayName:
        maxLength: 128
        type: string
      role:
        maxLength: 64
        type: string
      status:
        maxLength: 32
        type: string
      tenantId:
        type: integer
    required:
    - displayName
    - role
    type: object
info:
  contact: {}
  description: Clear Bill backend service API.
  title: Clear Bill API
  version: 0.1.0
paths:
...
  /api/v1/users:
    get:
      description: 超管可查询所有用户，租户管理员仅可查询本租户用户
      operationId: users-list
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 租户ID
        in: query
        name: tenantId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 执行成功
          schema:
            $ref: '#/definitions/clearbill_mgr_server_api_vo.ResponseResult'
      summary: 查询用户列表
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 仅超管和租户管理员可创建用户，默认密码为 bill123;
      operationId: users-create
      parameters:
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clearbill_mgr_server_api_vo.CreateUserReq'
      produces:
      - application/json
      responses:
        "201":
          description: 执行成功
          schema:
            $ref: '#/definitions/clearbill_mgr_server_api_vo.ResponseResult'
      summary: 创建用户
      tags:
      - user
  /api/v1/users/{id}:
    delete:
      description: 删除指定用户
      operationId: users-delete
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 执行成功
          schema:
            $ref: '#/definitions/clearbill_mgr_server_api_vo.ResponseResult'
      summary: 删除用户
      tags:
      - user
    get:
      description: 查询指定用户详情
      operationId: users-get
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 执行成功
          schema:
            $ref: '#/definitions/clearbill_mgr_server_api_vo.ResponseResult'
      summary: 查询用户详情
      tags:
      - user
    put:
      consumes:
      - application/json
      description: 更新指定用户信息
      operationId: users-update
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clearbill_mgr_server_api_vo.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 执行成功
          schema:
            $ref: '#/definitions/clearbill_mgr_server_api_vo.ResponseResult'
      summary: 更新用户
      tags:
      - user
  /api/v1/users/{id}/password:
    put:
      consumes:
      - application/json
      description: 超管可修改所有用户密码，租户管理员可修改本租户所有用户密码
      operationId: users-password-reset
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clearbill_mgr_server_api_vo.ResetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 执行成功
          schema:
            $ref: '#/definitions/clearbill_mgr_server_api_vo.ResponseResult'
      summary: 重置用户密码
      tags:
      - user
swagger: "2.0"